+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-15 20:16:06
-// Commit: feat(state): integrate offline caching for state state
+// Updated at: 2026-05-15 21:09:29
+// Commit: feat(state): integrate custom telemetry logger in state
 
 class stateComponentState {
-  final String commitRef = '1193';
-  final String lastMessage = 'feat(state): integrate offline caching for state state';
-  final DateTime timestamp = DateTime.parse('2026-05-15T20:16:06+05:00');
+  final String commitRef = '1197';
+  final String lastMessage = 'feat(state): integrate custom telemetry logger in state';
+  final DateTime timestamp = DateTime.parse('2026-05-15T21:09:29+05:00');
 }